YOU filter menu added, YOU actions mnd view menus changed
svn path=/trunk/ncurses-pkg/; revision=46728

+++ libyui-ncurses-pkg/src/NCPkgMenuAction.cc
@@ -38,34 +38,52 @@ NCPkgMenuAction::~NCPkgMenuAction()
 
 void NCPkgMenuAction::createLayout()
 {
-    toggleItem =  new YMenuItem( _( "&Toggle    [SPACE]" ) );
-    items.push_back( toggleItem );
+    if ( !pkg->isYouMode() )
+    {
+	// Please note: add an appropriate number of whitespaces to get a well
+	// formated menu (the [ ]s should be in one column)
+	// begin: Actions menu, toggle the status of a package, e.g. change from installed to delete
+	toggleItem =  new YMenuItem( _( "Toggle    [SPACE]" ) );
+	installItem = new YMenuItem( _( "Install     [+]" ) );
+	deleteItem =  new YMenuItem( _( "Delete      [-]" ) );
+	updateItem =  new YMenuItem( _( "Update      [>]" ) );
+	tabooItem =   new YMenuItem( _( "Taboo       [!]" ) );
+	lockItem =    new YMenuItem( _( "Lock        [*]" ) );
+	// end: Actions menu, set status of all packages (title of a submenu)
+	allItem =     new YMenuItem( _( "All Listed Packages" ) );
 
-    installItem = new YMenuItem( _( "&Install      [+]" ) );
-    items.push_back( installItem );
+	items.push_back( toggleItem );
+	items.push_back( installItem );
+	items.push_back( deleteItem );
+	items.push_back( updateItem );
+	items.push_back( tabooItem );
+	items.push_back( lockItem );  
+	items.push_back( allItem );
 
-    deleteItem =  new YMenuItem( _( "&Delete      [-]" ) );
-    items.push_back( deleteItem );
+	// begin: submenu items actions concerning all packages
+	installAllItem = new YMenuItem( allItem, _( "Install All" ) );
+	deleteAllItem = new YMenuItem( allItem, _( "Delete All" ) );
+	keepAllItem = new YMenuItem( allItem,  _( "Keep All" ) );
+	updateAllItem = new YMenuItem( allItem, _( "Update All Unconditionally") );
+	// end: submenu items: actions concerning all packages
+	updateNewerItem = new YMenuItem( allItem,  _( "Update If Newer Version Available" ) );
 
-    updateItem =  new YMenuItem( _( "&Update      [>]" ) );
-    items.push_back( updateItem );
+	addItems( items );
+    }
+    else
+    {
+	toggleItem =  new YMenuItem( _( "Toggle    [SPACE]" ) );
+	installItem = new YMenuItem( _( "Install     [+]" ) );
+	deleteItem =  new YMenuItem( _( "Delete      [-]" ) );
+	updateItem =  new YMenuItem( _( "Update      [>]" ) );
+	
+	items.push_back( toggleItem );
+	items.push_back( installItem );
+	items.push_back( deleteItem );
+	items.push_back( updateItem );	
 
-    tabooItem =   new YMenuItem( _( "T&aboo       [!]" ) );
-    items.push_back( tabooItem );
-
-    lockItem =    new YMenuItem( _( "&Lock        [*]" ) );
-    items.push_back( lockItem );
-
-    allItem =     new YMenuItem( _( "All Listed &Packages" ) );
-    items.push_back( allItem );
-
-    installAllItem = new YMenuItem( allItem, _( "Install All" ) );
-    deleteAllItem = new YMenuItem( allItem, _( "Delete All" ) );
-    keepAllItem = new YMenuItem( allItem,  _( "Keep All" ) );
-    updateAllItem = new YMenuItem( allItem, _( "Update All Unconditionally") );
-    updateNewerItem = new YMenuItem( allItem,  _( "Update If Newer Version Available" ) );
-
-    addItems( items );
+	addItems( items );
+    }
 }
 
 bool NCPkgMenuAction::handleEvent ( const NCursesEvent & event)
